+++ src/ScottPlot/Settings.cs
@@ -260,7 +260,7 @@ namespace ScottPlot
         /// <summary>
         /// Automatically adjust X and Y axis limits of all axes to fit the data
         /// </summary>
-        public void AxisAutoAll(double horizontalMargin = .1, double verticalMargin = .1)
+        public void AxisAutoAll(double? horizontalMargin = null, double? verticalMargin = null)
         {
             AxisAutoAllX(horizontalMargin);
             AxisAutoAllY(verticalMargin);
@@ -269,7 +269,7 @@ namespace ScottPlot
         /// <summary>
         /// Automatically adjust axis limits for all axes which have not yet been set
         /// </summary>
-        public void AxisAutoUnsetAxes()
+        public void AxisAutoUnsetAxes(double? horizontalMargin = null, double? verticalMargin = null)
         {
             Axis[] unsetAxesX = HorizontalAxes
                 .Where(x => !x.Dims.HasBeenSet && x.Dims.IsNan)
@@ -287,12 +287,12 @@ namespace ScottPlot
 
             foreach (Axis xAxis in unsetAxesX)
             {
-                AxisAutoX(xAxis.AxisIndex);
+                AxisAutoX(xAxis.AxisIndex, horizontalMargin);
             }
 
             foreach (Axis yAxis in unsetAxesY)
             {
-                AxisAutoY(yAxis.AxisIndex);
+                AxisAutoY(yAxis.AxisIndex, verticalMargin);
             }
         }
 
@@ -353,7 +353,7 @@ namespace ScottPlot
         /// <summary>
         /// Automatically adjust X axis limits to fit the data
         /// </summary>
-        public void AxisAutoAllX(double margin = .1)
+        public void AxisAutoAllX(double? margin = null)
         {
             int[] xAxisIndexes = Axes.Where(x => x.IsHorizontal).Select(x => x.AxisIndex).Distinct().ToArray();
             foreach (int i in xAxisIndexes)
@@ -363,18 +363,17 @@ namespace ScottPlot
         /// <summary>
         /// Automatically adjust Y axis limits to fit the data
         /// </summary>
-        public void AxisAutoAllY(double margin = .1)
+        public void AxisAutoAllY(double? margin = null)
         {
             int[] yAxisIndexes = Axes.Where(x => x.IsVertical).Select(x => x.AxisIndex).Distinct().ToArray();
             foreach (int i in yAxisIndexes)
                 AxisAutoY(i, margin);
         }
 
-        public void AxisAutoX(int xAxisIndex, double margin = .1)
+        public void AxisAutoX(int xAxisIndex, double? margin = null)
         {
-            double min = double.NaN;
-            double max = double.NaN;
-            double zoomFrac = 1 - margin;
+            if (margin.HasValue)
+                MarginsX = margin.Value;
 
             var plottableLimits = Plottables.Where(x => x is IPlottable)
                                                .Select(x => (IPlottable)x)
@@ -383,6 +382,8 @@ namespace ScottPlot
                                                .Select(x => x.GetAxisLimits())
                                                .ToArray();
 
+            double min = double.NaN;
+            double max = double.NaN;
             foreach (var limits in plottableLimits)
             {
                 if (!double.IsNaN(limits.XMin))
@@ -396,14 +397,15 @@ namespace ScottPlot
 
             var xAxis = GetXAxis(xAxisIndex);
             xAxis.Dims.SetAxis(min, max);
+
+            double zoomFrac = 1 - MarginsX;
             xAxis.Dims.Zoom(zoomFrac);
         }
 
-        public void AxisAutoY(int yAxisIndex, double margin = .1)
+        public void AxisAutoY(int yAxisIndex, double? margin = null)
         {
-            double min = double.NaN;
-            double max = double.NaN;
-            double zoomFrac = 1 - margin;
+            if (margin.HasValue)
+                MarginsY = margin.Value;
 
             var plottableLimits = Plottables.Where(x => x is IPlottable)
                                                .Select(x => (IPlottable)x)
@@ -412,6 +414,8 @@ namespace ScottPlot
                                                .Select(x => x.GetAxisLimits())
                                                .ToArray();
 
+            double min = double.NaN;
+            double max = double.NaN;
             foreach (var limits in plottableLimits)
             {
                 if (!double.IsNaN(limits.YMin))
@@ -425,6 +429,8 @@ namespace ScottPlot
 
             var yAxis = GetYAxis(yAxisIndex);
             yAxis.Dims.SetAxis(min, max);
+
+            double zoomFrac = 1 - MarginsY;
             yAxis.Dims.Zoom(zoomFrac);
         }